lsfd: fix the way to print length of mmap area
Suggested-by: Anatoly Pugachev <matorola@gmail.com>.

+++ misc-utils/lsfd-file.c
@@ -337,7 +337,7 @@ static bool file_fill_column(struct proc *proc,
 		if (file->association != -ASSOC_SHM
 		    && file->association != -ASSOC_MEM)
 			return true;
-		xasprintf(&str, "%lu", get_map_length(file));
+		xasprintf(&str, "%ju", (uintmax_t)get_map_length(file));
 		break;
 	default:
 		return false;
